refactor javascript
- use typescript
- add type annotations
- separate code into modules
