+++ .eslintrc.yml
@@ -14,3 +14,4 @@ rules:
   "no-unused-vars": "off"
   "@typescript-eslint/no-unused-vars":
     - "error"
+  "@typescript-eslint/ban-ts-comment": "off"
