+++ .pre-commit-config.yaml
@@ -19,6 +19,8 @@ repos:
     rev: "v8.12.0"
     hooks:
       - id: eslint
+        files: \.[jt]sx?$ # *.js, *.jsx, *.ts and *.tsx
+        types: [file]
         additional_dependencies:
           - eslint
           - typescript
