+++ package.json
@@ -18,6 +18,8 @@
   "devDependencies": {
     "shelljs": "0.8.5",
     "svgdom": "0.1.10",
+    "ts-loader": "9.2.8",
+    "typescript": "4.6.3",
     "webpack": "5.72.0",
     "yaml": "2.0.0"
   }
