--- src/index.js
@@ -1,1 +1,0 @@
-import "./js/script.mjs";
