+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "compilerOptions": {
+    "outDir": "./dist/",
+    "noImplicitAny": true,
+    "module": "es6",
+    "target": "es6",
+    "strict": false,
+    "allowJs": true,
+    "moduleResolution": "node"
+  }
+}
